agents: timelines, sketches, scripts and storyboards namespaces run on the capability registry (PR 8 of tool-class retirement)
Ports the four creative-surface tool families, including
validate_timeline / validate_sketch whose loaders become
run.loaders.{timeline,sketch}. Existing surface suites pass unmodified
against the deprecated thin subclasses.
